Prevent erroneous migration from being generated
The choices used for the "jobpage" model "icon" field are generated by listing
the contents of a directory on the filesystem. Sort the items so the order is
consistent. The method used to list the items (os.listdir) returns entries in
an arbitrary order which can make it appear that the choices have changed and,
consequently, can trigger the creation of a new but erroneous migration.