Suppress test warnings
The behaviour of the @factory.post_generation decorator has changed which
results in a deprecation warning when we run our tests. See

https://factoryboy.readthedocs.io/en/latest/changelog.html#id1

> DjangoModelFactory will stop issuing a second call to save() on the created
> instance when Post-generation hooks return a value.

In our case, the call to the self.file.save method in the
@factory.post_generation decorated method ultimately calls the save method on
the created instance anyway so we can safely disable the post generation method
save and we don't get a bunch of crud in the output of our tests.