Set base height for header messages
At present, a header message is only used to display the office
closed notice and is conditionally displayed only when the office
is actually set as closed so we don't need to have the height set
to zero.

Also fix the position of the shutdown notice for a certain device
size so the notice still displays when the user scrolls.