Prevent tenancy serialization error
In the past the tenancy is_on_lease field was a boolean field. It was converted
to a char field and the existing boolean values were converted to text.
Subsequently choices were specified on the field but the existing values were
not mapped to the choice values. This change maps the old boolean text values
to the new choice values:

- "TRUE" maps to "YES"
- "FALSE" maps to "NO"

This prevents errors when serializing tenancy objects.