Add issue employment_status choice value back to prevent serialization error
Since we are not removing the existing value from the data we should keep the
existing choice value to prevent errors when serializing issue objects.

Added text in the human readable name part of some employment_status choice
values to indicate that they should not be used i.e. they are legacy values.