Fixed rendered email html using potentially outdated data
Prior to this change the html fragment passed to the function used to generate
the complete html content of the email was from the record before it was saved
rather than from the incoming data that would subsequently be used to update
the record. If the update included changed html then outdated data would be
used to generate the html email content.

The fix was to simply use the incoming html data to generate the complete html
email content.

Probably better to generate the full html content pre-save or on save of the
Email model instance. Then you don't have to rely on the update method being
called at all which is a bit more robust to future changes. But I have left it
as-is for now.