Fix Uncaught TypeError when there are no emails associated with a case
The backend method (EmailApiViewset:retrieve) that serves as the API endpoint
for retrieving email threads returns a 404 response if there are no email
threads associated with a case. We get an unhandled error

    Uncaught TypeError: Cannot read properties of undefined (reading 'length')

when we try to read emailThreads.data.length property to determine whether
there are any email threads associated with the case (because emailThreads.data
is undefined).

Because the 404 error here means there are no associated email threads, the fix
is to handle that error specifically to prevent the error and display the "No
emails found" message.