Fix python ImportError when building docker image
The full error is

  ImportError: cannot import name 'Sanitizer' from 'html_sanitizer' error

This is ultimately caused by an indirect dependancy on lxml. The lxml project
is a dependency of our direct dependency html-sanitizer. The lxml project split
the HTML cleaning portion of the project, which we use, into a separate
project. See here for details:

https://lxml.de/5.2/changes-5.2.0.html

Presumably an implicit dependency on the new project, lxml_html_clean, is
included as part of an updated version of html-sanitizer but the version we
used is pinned. The fix applied is to explicitly install lxml & lxml_html_clean.