Simplify test debugging
Using the debug option of the invoke "test" command we can debug tests in
vscode similarly to debugging clerk. For example, run the tests with debugging
active:

  $ inv test --debug
  [+] Creating 1/0
   ✔ Container clerk_db  Running
  Waiting for debug client to attach...

and then use vscode to attach to the debugger using the following launch
profile:

  {
    "version": "0.2.0",
    "configurations": [
      {
        "name": "Run debugger",
        "type": "python",
        "request": "attach",
        "django": true,
        "justMyCode": false,
        "pathMappings": [
          {
            "localRoot": "${workspaceFolder}/app",
            "remoteRoot": "/app"
          }
        ],
        "port": 8123,
        "host": "127.0.0.1",
      }
    ]
  }

This is the same launch profile used when debugging clerk.