Resolve task assignee when the task is created
Figure out who the task should be assigned to when we create the task rather
than, as previously considered, using the task role to figure out the user
whenever it was required. The down side of this change is that we have to add a
bunch of special cases to handle changes to case assignment (i.e. the assigned
paralegal or lawyer). But to handle ad hoc tasks we need to be able to assign
to a specific user and we don't want ad hoc & automatically created tasks to
have different ways of specifying who the task is assigned to.