Fix use of incorrect logger
If we adjust the log output format to include the log name we see that, prior
to this change, the log name for some messages was "root". We want the correct
(package hierarchy) name (see
https://docs.python.org/3/library/logging.html#logger-objects) to be included
in the output so we have a better idea where a log message originates.

This looks to simply be an incorrect use of the root logger rather than the
module level one. To address this we just change to use the module level
logger.