Allow specifying specific fields to only be included in the task API response
A response including a serialized issue object is very heavyweight especially
since we typically don't need most of the fields. Response time, particularly
when we are returning a list of tasks, is considerably improved if we restrict
the fields to only those that we need.

I have used an existing library for this so as not to have to specify multiple
different serializers by hand.