Use months instead of minutes when generating dates for test data
The faker library documentation for methods that generate dates (like
date_time_between) states that it:

  Accepts date strings that can be recognized by strtotime().

Unfortunately it's not clear where this function is described. At any rate, it
seems like "m" specifies minutes and "M" specifies months.

So, for example, when specifying the times for the date_time_between faker
method, a start value of "-2m" and end value of "-1m" specifies a time between
1 and 2 minutes ago.

For the created and modified dates for objects used in tests, values in that
range seem pointless so I have changed to use months instead of minutes.