Support TableForm component display value & choices not inline with API response
At present some Clerk API field responses include the actual (i.e. from the db)
value but also the "display" value and the possible choices that the field
allows. This results in a lot of duplication for API list responses and means
that the data you get and the data you post back to the API are quite
different.

To address these issues, this change supports display values and choices for
the TableForm component that are not supplied inline with the API response. The
component still supports inline values so we can change usage over time.