Use full standards mode document layout by supplying a DOCTYPE
Unless there is a good reason not to use full standards mode - and there
doesn't seem to be - then we should as it's easier to reason about layout when
it follows the current standard.

This change also gets rid of browser console warnings about using quirks more
e.g. this one from firefox:

  This page is in Quirks Mode. Page layout may be impacted. For Standards Mode
  use “<!DOCTYPE html>”.