Prevent duplicate log messages
The django specific logger propagates log records to the root logger which has
exactly the same handlers, this causes each log record to be duplicated. We
could have set the django-specific logger's propogate attribute to "False" but,
given that the django & root logger's attibutes are the same it seemed
pointless to keep the django logger.

For more details see

- https://stackoverflow.com/questions/62882552/duplicate-log-output-when-using-django-logging-module
- https://docs.python.org/3/library/logging.html#logging.Logger.propagate