Add guard condition to prevent intermittent KeyError
There are some circumstances where the context dictionary does not contain a
"request" item. In those cases we get a KeyError, for example:

  KeyError /clerk/api/email/{pk}/{email_id}/attachment/{attachment_id}/sharepoint/
  Level: Error
  Unhandled 'request'

Those circumstances are unclear however. It seems to be related to the issue
that was addressed in commit da954958c6713837d23e4098e952d769ab24db8a but I
cannot reproduce it. This change seems like a reasonable precaution until we
have more information.