Pin lxml version to prevent ImportError
Full error is

  ImportError: cannot import name 'Sanitizer' from 'html_sanitizer'

Note lxml is a dependency of html-sanitizer.

Not quite sure what the issue is here but a potentially better fix would be to
update html-sanitizer. I don't have the time to test changes to this package
now and a project to update all dependencies including testing is already
planned. Pinning the lxml version will do for now.