Simplify creating tasks for testing
Assign the owner of a factory created task as the paralegal of the associated
issue. This simplifies creating tasks for testing as, otherwise, paralegal
users do not have access to view tasks where they are not also assigned as the
paralegal of the issue.

Update some tests that implicitly relied on the issue not having an assigned
paralegal.