Change to use uv instead of pip for Python package management
The primary reason for doing this is lockfile support but we also get

- Nice docker integration
- Built-in venv support
- Package install speed boost
- Python executable version management
- Better documentation

See https://docs.astral.sh/uv/ for documentation.

A big change introduced here is that we let uv take care of installing the
python executable rather than installing it in the base docker image. Strictly
speaking we don't need to do this but it makes debugging dependencies running
on docker (i.e. not our code) much easier because we can install the same
python version (using uv) locally.