Update to django version > 5 and address deprecations
Because we want to get the security updates & other features in later django
versions.

We have to change the django-q dependency to django-q2 as development
of django-q has stalled and it has not be updated to support django version >
5. Without this change we get the following error that ultimately originates
from django-q:

  ImportError: cannot import name 'baseconv' from 'django.utils'

as the baseconv module has been deprecated in django 5. See
https://code.djangoproject.com/ticket/32712

The django-q2 package is a drop-in replacement for django-q with ongoing
development.