Fix social-auth-app-django dependency data model issue
Fix an issue when running a migration for the social-auth-app-django
(social_django) package after updating. It's not clear why this happens,
possibly we messed up a previous update?

The issue is triggered here:

  File "/app/.venv/lib/python3.12/site-packages/social_django/migrations/0013_migrate_extra_data.py", line 14, in migrate_json_field
    UserSocialAuth.objects.using(db_alias).exclude(extra_data='""').iterator()

And results in the following error:

  django.db.utils.ProgrammingError: operator does not exist: text = jsonb
  LINE 1: ...RE NOT ("social_auth_usersocialauth"."extra_data" = '"\"\""'...

The fix is to change the extra_data field data type from text to jsonb. For
further details see
https://github.com/python-social-auth/social-app-django/issues/209.