Change method used to stop delete service click erroneously triggering callback
Preventing propagation is less esoteric that checking if an element is
connected. Also it lives outside of the hook used to check for "outside clicks"
so it potentially gives that hook more flexibility, although I can't think of a
circumstance where you would want to check for clicks outside of a disconnected
element.