Remove cancel button when confirming service delete & adjust click outside handling
We don't really need the cancel button as you can easily dismiss the delete
confirm button by clicking anywhere else. It seems redundant.

Use CSS to dynamically display the service action buttons or the delete
confirmation instead of returning only the elements we want to display (i.e.
the react way) to get the click outside functionality that we want.

If we, for example, dynamically return only the elements we want to display
then in some circumstances the click that adds the click event listener also
triggers it and as the event target is no longer connected to the document it
is not contained within the element (i.e. it appears to be a click outside of
the element). We can detect if the target is not connected but sometime we want
to trigger the outside click event handler in that case so we can't use it for
differentiation. If we stop propagation of the initial click event then we
don't recognise a click outside when clicking the delete icon on another row.