Correct case validation error to be non-field specific
Previously we would report a validation error for unfinished ongoing services
when closing a case to be specific to the "stage" field. This isn't strictly
correct as the issue is with a missing field of a related service and changing
the stage field could not correct the error. Given that we still want to
perform this check and the error is related to the field of a different model
we now report it as not related to a specific field.