Handle ordering of service issue notes when a failure occurs
Prior to this change if the creation of a service issue note fails and is
processed again at a later time it would have the creation date of when it was
processed not when (roughly speaking) the service was created, updated or
deleted. This would mean that its order in the timeline would potentially be
incorrect.

By supplying a datetime when the task is created the issue note should be
ordered correctly in the timeline irrespective of when it is processed.