Lower the log level when notification are without a target for cases closed unassigned & unstarted
Cases are commonly closed unassigned (i.e. without an assigned paralegal or
lawyer) & unstarted (i.e. the previous case stage is "UNSTARTED") for a variety
of valid reasons. See https://linear.app/anika-legal/issue/TEC-1387/ for more
details.

Prior to this change we would log an error when users closed a case unassigned
& unstarted which would ultimately trigger the creation of an issue in our
error monitoring system (Sentry).

In these cases we now log a message at the info level as this is a common &
valid action that does not represent an error.