Guard against AttributeError when checking API permissions
The request user instance is augmented in middleware with some attributes that
specify Clerk permission settings. The Django Rest Framework (DRF) clones the
request object and overwrites the request user instance for unauthenticated
users with an new anonymous user object so these attributes are lost (e.g. see
DRF request.py github link below). This results in an AttributeError if we try
to access these attributes. See Sentry issue CLERK-2QF at
https://anika-legal.sentry.io/issues/5418662710/.

An example of a specific error of this type is:

  AttributeError 'AnonymousUser' object has no attribute 'is_coordinator_or_better'

I can think of a couple of ways to handle this and have chosen the easiest of
those options; adding a guard condition to check if the user is authenticated
before trying to access the permission attributes.

An alternative could be to specify the DRF UNAUTHENTICATED_USER setting (see
https://www.django-rest-framework.org/api-guide/settings/#unauthenticated_user)
to our own AnonymousUser (django.contrib.auth.models.AnonymousUser) subclass
that include the permission attributes (set to False). But it would be a bunch
more work to set this up so that the attributes were not defined in multiple
places.

- DRF request.py: https://github.com/encode/django-rest-framework/blob/9016efe3fc412488df92912c619f8f24fed2937c/rest_framework/request.py#L406).