Make all email attachments reference the same file in the staging env as a form of obfuscation
We don't want to reference the actual email attachments in the test system as
they may contain PII or other sensitive data. We could generate fake files
which would be nice for testing but this seems impractical - it woulds take too
long to generate and upload files via this script so it would have to be done
via another task. For the moment we will just reference the same,
non-sensitive, file for all attachments.