Maximise the query parameter expiry of generated AWS S3 URLs for file uploads
From a strict security standpoint this is not ideal but it is a pragmatic way
to handle the URLs from intake form file uploads not expiring if a user does
not complete their submission in a timely fashion.

The issue that this addresses is minor as only the text link in the intake form
becomes invalid after the expiry period but the file upload itself is still
correctly associated with the submission and, if the submission is completed,
the created case.