Don't redirect to S3 when serving wagtail (blog) documents
This handles a possible situation where users hold on to a document URL (e.g.
bookmark or leaving tab open) that would expire in, at the time of writing,
one hour due to the use of AWS S3 signed URLs.

Instead links to documents point to a Django view which both checks the user’s
permission and serves the document. See
https://docs.wagtail.org/en/stable/reference/settings.html#wagtaildocs-serve-method