Add slack notification when email delivery fails
If the email is related to a case with an assigned paralegal then the Slack
notification goes to the paralegal otherwise it goes to the alerts Slack
channel.

Before this change there wasn't good visibility when email delivery failed.
This situation was being managed in an ad hoc manner when it was noticed via
the email UI. This change will hopefully allow for more proactive management of
delivery failures.