Prevent clerk login before invitation from creating an account
Prior to this change, when a user with an anikalegal.com Google Workspace
account logs into Clerk and they do not already have a local Clerk account then
an account is automatically created for them. The account is assigned default
permissions to edit the Clerk CMS (Wagtail).

We want to change this so that no one can login before a local Clerk account
has been created for them as paralegals have a tendency to do this which means
they are assigned permissions that they should not have and this ultimately
requires more administration work for us.

This change removes the auth setup that automatically creates an account and
assigns it default permissions. It also displays a message on the login page
when a user without a local Clerk account tries to login.

See https://linear.app/anika-legal/issue/TEC-1413/