Add files to storage to use for obfuscation
We don't want to copy email attachments & file uploads from production to
staging environment storage (AWS S3) for privacy & security reasons so we have
to obfuscate references to files in storage to point to another file.

Previously we were relying on an existing file just being present in staging
file storage but that is unreliable so we add our own sample files.