Add some asset directories to sync when updating staging with production data
The added directories are:

- action-documents: previously in use but just missing from the sync.
- documents: new.

Also remove asset sync task from command line runner script. We shouldn't sync
assets without also updating the data as we're likely to delete files we need
and sync files that are not used. Also the code is likely to get out of sync if
it is defined in multiple places. Use the GitHub Staging workflow to update the
data AND sync storage assets.