Change backup so AWS access keys don't have to be accessible on the prod server
We stream the database backup from the server over SSH and copy directly to AWS
S3. This means we don't have to have the access keys on the prod server. The
ultimate goal is to remove backup bucket write access from any user whose creds
are accessible on the prod server. Thus if the prod server is compromised the
online backups are still secure.

This change requires that we configure our GitHub actions secrets and variables
with the access keys of a user configured to have read/write access to the
backup s3 buckets.

Also note that we don't bother piping the database dump to gzip as this is
handled by pg_dump itself:

  For the custom and directory archive formats, this specifies compression of
  individual table-data segments, and the default is to compress using gzip at
  a moderate level.

-- https://www.postgresql.org/docs/current/app-pgdump.html