Update wiki connection and S3 key info
The clerk.pem private key has been lost so we couldn't connect to the AWS EC2
instance that hosts the wiki without setting up a new OpenSSH key. The IP
address also changed as we have to stop & restart the instance as part of
setting up a new key.

We also restrict access to necessary S3 buckets only by setting up an AWS user
with permission policies that only allow access to specific S3 buckets &
assigning them access keys. The keys are configured elsewhere but we note the
full access key id here as it's not secret and it helps us see where the key is
being used e.g. if we want to rotate it.