Trial simplifying running docker commands in GitHub actions using docker context
This is a trial change to the script that deploys the clerk application to use
docker context instead of manually copying docker files from the GitHub action
runner to our staging/prod server.

If we keep this then a bunch of changes to the GitHub action workflow files &
the commands that are executed by them will be required.

The intention is to simplify the code and make running commands on docker
easier. For example, running a Django management command to export client info
for disaster recovery. See https://linear.app/anika-legal/issue/TEC-1423/