Use the previous method to reset the staging database due to perms issues
The django reset_db management command ultimately runs SQL commands on the
postgres db as the clerk user within the docker container but the clerk user
doesn't have permissions to create any databases so the reset fails. I don't
want to change this without reviewing the implications so I'll just revert this
to the previous method for the moment.