Use client-side specified display text for service category & type choices
We enumerate the possible category & type values in the OpenAPI schema and use
the generated types to specify constants for the display text.

The only benefit of sending the display text as "context" from the backend that
I can see is that if we add to the Django choices then the new value is
available automatically on the frontend. This is a great benefit but it comes
at the cost of having to shuffle around untyped or poorly typed data on the
frontend. If we make a change (rather than an addition) on the backend then we
likely also have to change the frontend without the benefit of decent typing.

The benefit of simpler frontend code with better typing seems worth the trade
off.