Fix frontend TypeError when editing tenancy with null value for is_on_lease field
The full frontend error is

  TypeError: Cannot read properties of undefined (reading 'map')
    at field-component.tsx:109:26

Some choice fields, like the tenancy is_on_lease field, include the field
value, the field's display value & a list of the possible choices that the
field allows when serialized. The mechanism that generates these values (the
TextChoiceField.to_representation method, see link below) is not called
automatically when the field value is null. The results in a TypeError when the
frontend attempts to access those field values as the serialized value is null
instead of the expected structure.

The simplest way to correct this without changing the existing code is to set
the empty string as the default for the tenancy model is_on_lease field. The
TextChoiceField.to_representation method is then called automatically.

References:

- TextChoiceField.to_representation:
  https://github.com/AnikaLegal/clerk/blob/f3baf6fd95fd6bebc48fa8f6b693f99d2e303eda/app/case/serializers/fields.py#L10
- https://stackoverflow.com/questions/54257477/custom-django-rest-framework-serializer-field-not-running-to-representation