Update GitHub workflows to allow releasing any git branch to the staging environment
This allows us to get changes, particularly those that may take a long time to
approve for production, in front of people without having to merge them to
staging first.

This avoids the difficult & error prone process of having to manually cherry
pick approved changes for production while other changes are languishing on the
develop branch waiting for approval. This is mostly relevant to large projects
that can take some time for others to review & produce or update documentation
etc.