Wait on the result of processing audit logs so we can update the frontend UI
We could just run the task synchronously but this makes it easier to rerun the
task if it fails and allows us to track successes & failures without having to
trawl log files.

Also added a return value for functions scheduled to run on the django q task
queue. This is only necessary if you have to wait on the result of a task. If
not specified then the full wait time is used rather than just waiting until
the task is finished.