fix(core): change when we adjust document perms & case assignment notification
Previously we adjusted access to documents associated with a case and notified
users of case assignment prior to the case being saved (i.e. when we got the
Issue model pre_save signal). This is problematic as if the save subsequently
failed we would have changed the perms in an external system and sent a
notification that did not reflect our internal state.

We can instead handle this situation after an issue event save (i.e. when we
get the IssueEvent model post_save signal) as:

- The event is created under the same circumstances that we tested for
  previously (i.e. when the assigned paralegal changes).
- The event contains all of the details that we need for the required tasks.
- It addresses the issue mentioned above.
- It is (to me) aesthetically cleaner.

Some adjustment to the paths of patched functions or adding patched functions
for tests was also required. No changes were required to the test code itself.