fix(emails): fix SuspiciousFileOperation error saving inbound email from SendGrid
An example of the error in full is:

  SuspiciousFileOperation

  Storage can not find an available filename for
  "email-attachments/e1d245e34b686548e1bf6f02b9842c10/h_l51h1QP.application-to-reopen-made-by-another-agency.pdf".
  Please make sure that the corresponding file field allows sufficient
  "max_length".

This is addressed by increasing the maximum allowed file name length for
attachments in the database.

By default, the maximum length for Django FileField fields is 100. We are using
AWS S3 as our storage backend and the file name includes the S3 bucket name and
a hash of the file contents (used as a "folder" in S3).  Effectively this
limits the file name to ~50 bytes. Looking through our logs I can see that this
issue occurs semi-regularly so this value is obviously too low for the file
names that our users are attaching.

I have increased it to allow the maximum file name length that S3 can handle.
See https://docs.aws.amazon.com/AmazonS3/latest/userguide/object-keys.html.