fix(emails): suppress duplicate email delivery failure notifications
The issue was caused because we were handling the "processed" SendGrid email
event type but not doing anything other than saving the unchanged email db
instance again. This could lead to a race condition when we got "processed" &
"bounce" (or "dropped") events in a short time span which I think goes
something like this:

1. Send email
2. Receive "processed" event
3. Receive "bounce" event
4. The race happens here.

If the "processed" event is handled first the issue does not occur.

If the "bounce" event is handled first then we set the state to delivery
failure and (ultimately, see save_email function link below) queue a
notification. As part of handling the notification, we set a flag on the email
record to indicate that the notification has been sent. While the queue is
being processed we handle the "processed" event, save the email record and,
because the email state is set to delivery failure, ultimately trigger the
notification again as we haven't set the notification flag yet.

The fix here is to just drop the handling of the "processed" event as it wasn't
being used to indicate any change to the email state.

I also refactored the code a lot for aesthetics & added some locking to handle
potential duplicate SendGrid events (see link below). This isn't the ideal way
to handle duplicate events but other possible options (e.g. saving events to a
database with a unique field for the event sg_event_id value and processing
them from there) are too much work for the potential benefit.

References:

- save_email: https://github.com/AnikaLegal/clerk/blob/b90f28196f5f04c2a6f6be369edf25d276b6828a/app/emails/signals.py#L11
- https://www.twilio.com/docs/sendgrid/for-developers/tracking-events/twilio-sendgrid-event-webhook-overview#duplicate-events).