feat: add empty choice to serializer text choice fields if the field allows empty values
This is a convenience as we could do it e.g. in the frontend but then we would
have to do it in multiple locations whenever it was required.

As a consequence of this change we also update the tenancy is_on_lease
serializer field to match the existing frontend behaviour i.e. an empty value
is not allowed.