fix(emails): NoReverseMatch: Reverse for 'case-email-thread' with arguments ... not found
The error is triggered when viewing case emails where an email has a subject
that returns the empty string when slugified.

We thread emails by subject and use the slugified subject as the URL for
viewing the email thread. As the Django slugify function will remove characters
that are not allowed in the path part of a URL it is possible to generate an
empty string as the URL for some subjects. This causes the error when we
attempt to generate the full URL for that email thread.

Threading emails by subject and using the subject as part of the URL for the
thread are, in my opinion, wrong. Primarily because emails can legitimately
have no subject and emails with the same subject can belong to different
"conversations". There isn't scope to fix this here as it would involve a lot
of breaking changes.

The fix applied is to use a URL of "no-subject" when we cannot generate one.
This is not a good fix but I think the right one for now.

See
- https://linear.app/anika-legal/issue/TEC-1523/
- https://docs.djangoproject.com/en/5.1/ref/utils/#django.utils.text.slugify