feat(web): add Wagtail snippet to display call-to-action banner on website
Allows CMS users to create a new banner and specify the banner properties
(title, subtitle, image etc.) and which banner is currently active. We make
sure that only a single banner can be active at one time.

This replaces some ad hoc code that I would change every time we needed a new
banner. The goal was to make this easier and faster for myself and possibly
allow others to do this as I was having to update the banner quite often.