refactor(microsoft): generate document template folder path from the case topic
Instead of hardcoding file folder path we generate it from the supplied case
topic. This means we don't have to remember to add a new path whenever we add a
new topic.

Also update some existing template folder names to match the value returned by
the function (Django slugify) we are using to generate the name from the topic
type.