fix: allow case closure when there is an associated but deleted ongoing service
For reasons that are not relevant here, services are soft-deleted i.e. they are
marked as deleted but not removed from the database. Business logic prevents
closing cases with associated unfinished ongoing services. The issue here was
that deleted unfinished ongoing services were preventing case closures which is
not reasonable or correct.

Add some tests to formalise the functionality we require.