refactor: remove prefix from issue-specific intake form answers
The prefix comes from values collected in the intake form and the use of
prefixes there is an implementation detail. Keeping this in clerk is
unnecessary and buggy:

- The code to strip the prefix on the front-end didn't work for issues with the
  RENT_REDUCTION topic (or any multi-word topic).

- The prefix used for evictions for arrears (i.e. "EVICTIONS") doesn't match
  the issue topic (i.e. "EVICTION"). This worked but made adding the new
  retaliatory evictions topic problematic.

We remove the prefixes from existing records and strip them when processing
intake form submissions. Some intake form adjustments were necessary to allow
using the topic value to strip the prefix when processing submissions.

We also remove the front-end code to strip the prefixes when displaying them as
it is now unnecessary.