feat(frontend): don't create client or tenancies separately from issues
Previously the functionality was such that you could only create clients or
tenancies before creating the issue itself. This opened the possibility of
creating clients or tenancies that were not associated with an issue. This
change prevents that possibility as it is not valid to have a client or tenancy
that is not associated with an issue.