fix(api): handle exception getting template mod/created time from msgraph
We want to be able to have a finer-grained approach to dealing with errors
retrieving details from the Django storage API (backed by MSGraph in this
case). For example, if we are listing the templates we don't want an issue
retrieving one of the files to prevent listing the other files.

See https://linear.app/anika-legal/issue/TEC-1557/