feat(microsoft): add url quoting & conflict handling when uploading files via msgraph
The previous msgraph upload file code did a bunch of inefficient & difficult to
understand stuff that I *think* was an attempt to handle file names with
characters that are not allowed in a URL. I simplified this by using URL
quoting and also adding support for the msgraph conflictBehavior attribute. We
default to fail when there is a file name conflict so we don't accidentally
overwrite existing files.

See:
- https://learn.microsoft.com/en-us/graph/api/resources/driveitem?view=graph-rest-1.0#instance-attributes
- https://docs.python.org/3/library/urllib.parse.html#urllib.parse.quote