fix(core): replace document template name field with annotation
We already have the file name as part of the file field so if we use an
annotation we don't risk the two fields getting out of sync.

Also setting the name value before calling the parent save method didn't work
properly as the file field name can change during save. This could be fixed by
setting it after save & saving again but that is unnecessary if we use an
annotation.