refactor(microsoft): remove concept of base path from msgraph django file storage
I included this concept to "restrict" files to being placed under a specific
location but the value of this doesn't seem worth the added code complexity and
opaqueness. N.B. files are still restricted to the location specified by their
upload path which now included the "base path". I think this made more sense
when I was considering using ids to access files in msgraph.